Square highlighting x-ray moves
If a square is clicked for details, all pieces that can attack the square are shown. This now includes pieces that can attack by x-ray, only if the piece that is blocking them can also attack the square. An example would be a rook battery can each attack the same squares along one axis, this would now be shown